Optimizes automatic refresh behaviour
Stop the timer while the application is not visible. Restart the timer once the
application becomes visible. This fixes the issue that QML timer slow down and
finally stop when the device enters sleep mode. This should also reduce battery
drain.

Set a timestamp with every refresh. Do not set the timer to trigger on automatic
refresh interval. Trigger every 5 minutes instead and check the refresh
timestamp to avoid automatic refreshes right after a manual refresh.

The five minute resolution of the automatic refresh timer is tolerable.